backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.5.3",
    "django>=5.2.6",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.15.2",
    "psycopg2-binary>=2.9.10",
    "redis>=6.4.0",
]
